requirements.txt
Flask
python-wordpress-xmlrpc
requests
